Add compatible OpenAI model types
The current model enum and methods that have models as parameters could easily cause errors as any model could be chosen for any of those methods. By adding the new `OpenAIEndpointModel` type that corresponds to the compatibility list in the OpenAI docs, it’s much easier to select the fitting model. The new type does contain duplicate model strings, which is intended to make it easier to maintain. Backward compatibility is assured by marking the previous methods as deprecated. https://platform.openai.com/docs/models/model-endpoint-compatibility